requirements.txt
flask
flask-wtf
email_validator
flask-sqlalchemy
flask-bcrypt
flask-login
pyjwt
Pillow